requirements.txt
pyopengl>=3.1.6
pyopengl-accelerate>=3.1.6
numpy>=1.24.0
pyqt6>=6.5.0
pyglm>=2.7.0

